refactor: optimize group selector
